pyproject.toml
[tool.poetry]
name = "serial-monitor"
version = "0.1.0"
description = ""
authors = ["asdasdfasf <asdasdfasf@afas.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyqt6 = "^6.7.0"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
